Add regression tests for MCP feature
