+++ e2etests/tests/mcp/mod.rs
@@ -1,1 +1,2 @@
 pub mod test_mcp_command;
+pub mod test_mcp_command_regression;
